requirements.txt
# Core ML/CV libraries
ultralytics>=8.3.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Web apps (optional)
gradio>=4.0.0
streamlit>=1.28.0

# Labeling tools integration
roboflow>=1.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
ipykernel>=6.25.0
jupyter>=1.0.0


-e .